[List]Seperate `List` into two list's
A `SourcesList` and `TopHeadlinesList`, the data that should be rendered is very different, the `List` componant will be a mess and will have a lot of conditions.
Also, we pass the data from `SourcesScreen` to `SourcesList`, while what will be done in `TopHeadlinesList` is fetching the data from it directly.