[historySlice]Remove idNumbers
The data we get from the API doesn't have a unique `id` for each element, so my initial though was to give each elemnt we a store a unique id so that we can update the history as we need.
It's true that there is no `id` in the data we get, but we have `url` which the direct url to the articles, and that what we will us as  `id`.